copilot/requirements.txt
# Master_Brain Agent Python Dependencies
# Install with: pip install -r requirements.txt

# Database connectivity (A1: Relational, A2: Memory)
psycopg2-binary>=2.9.0
